refactor: updated directory structure and got rid of unused code
