+++ frontend/package.json
@@ -20,6 +20,7 @@
     "eslint": "5.12.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-error-boundary": "^4.0.4",
     "react-router-dom": "^6.11.2",
     "react-scripts": "^5.0.1",
     "recharts": "^2.6.2",
@@ -46,7 +47,8 @@
   },
   "devDependencies": {
     "@babel/core": "^7.22.1",
-    "@babel/preset-env": "^7.22.1"
+    "@babel/preset-env": "^7.22.1",
+    "prettier": "^2.8.3"
   },
   "main": "index.js",
   "license": "MIT"
