+++ frontend/src/App.test.tsx
@@ -1,9 +1,13 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
 import App from './App';
+import { MemoryRouter } from 'react-router-dom';
 
 test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
+  render(
+    <MemoryRouter>
+      <App />
+    </MemoryRouter>,
+  );
+  const linkElement = screen.getByText(/Runners/i);
   expect(linkElement).toBeInTheDocument();
 });
