+++ frontend/src/NavigationDrawer.tsx
@@ -1,188 +1,122 @@
 import * as React from 'react';
-import { styled, createTheme, ThemeProvider } from '@mui/material/styles';
-import CssBaseline from '@mui/material/CssBaseline';
-import MuiDrawer from '@mui/material/Drawer';
-import Box from '@mui/material/Box';
-import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
-import Typography from '@mui/material/Typography';
-import Divider from '@mui/material/Divider';
+import { Route, Routes, useNavigate } from 'react-router-dom';
+import {
+  AppBar,
+  Drawer,
+  ListItemButton,
+  ListItemIcon,
+  ListItemText,
+  useMediaQuery,
+  Grid,
+  Link,
+  Divider,
+  Typography,
+  List,
+  Toolbar,
+} from '@mui/material';
+import { useTheme } from '@mui/material/styles';
 import IconButton from '@mui/material/IconButton';
-import Badge from '@mui/material/Badge';
-import Container from '@mui/material/Container';
-import Grid from '@mui/material/Grid';
-import Paper from '@mui/material/Paper';
-import Link from '@mui/material/Link';
 import MenuIcon from '@mui/icons-material/Menu';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import NotificationsIcon from '@mui/icons-material/Notifications';
-import Timer10SelectIcon from '@mui/icons-material/Timer10Select';
-import { mainListItems, secondaryListItems } from './listItems';
-import Chart from './Chart';
-import { QueueSize, QueueData } from './Queue/Queue';
-import { ProgressSize, ProgressData } from './Progress/Progress';
-import Completed from './Completed';
-import Workflows from './Workflows//Workflows';
-import Runners from './Runners';
-import GitHubStatus from './components/GitHubStatus';
-import { useEffect, useState } from "react";
-import { Route, Routes } from 'react-router-dom';
-import Dashboard from './Dashboard';
-
-function Copyright(props: any) {
-  return (
-    <Typography variant="body2" color="text.secondary" align="center" {...props}>
-      {'Copyleft © '}
-      <Link color="inherit" href="https://github.com/Greuel">
-        Greuel
-      </Link>{' '}
-      {new Date().getFullYear()}
-      {'.'}
-    </Typography>
-  );
-}
-
-const drawerWidth: number = 240;
-
-interface AppBarProps extends MuiAppBarProps {
-  open?: boolean;
-}
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== 'open',
-})<AppBarProps>(({ theme, open }) => ({
-  zIndex: theme.zIndex.drawer + 1,
-  transition: theme.transitions.create(['width', 'margin'], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
-
-const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
-  ({ theme, open }) => ({
-    '& .MuiDrawer-paper': {
-      position: 'relative',
-      whiteSpace: 'nowrap',
-      width: drawerWidth,
-      transition: theme.transitions.create('width', {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.enteringScreen,
-      }),
-      boxSizing: 'border-box',
-      ...(!open && {
-        overflowX: 'hidden',
-        transition: theme.transitions.create('width', {
-          easing: theme.transitions.easing.sharp,
-          duration: theme.transitions.duration.leavingScreen,
-        }),
-        width: theme.spacing(7),
-        [theme.breakpoints.up('sm')]: {
-          width: theme.spacing(9),
-        },
-      }),
-    },
-  }),
-);
-
-const mdTheme = createTheme();
+import DashboardIcon from '@mui/icons-material/Dashboard';
+import AltRouteIcon from '@mui/icons-material/AltRoute';
+import Dashboard from './Pages/Dashboard/Dashboard';
 
 function NavigationDrawer() {
-  const [open, setOpen] = React.useState(true);
-  const [drawerPadding, setDrawerPadding] = useState(0);
+  const navigate = useNavigate();
+  const theme = useTheme();
+  const [menuOpen, setMenuOpen] = React.useState(true);
+  const [drawerPadding, setDrawerPadding] = React.useState(0);
+  const isMobile = useMediaQuery(theme.breakpoints.down('md'));
+
   const measuredDrawerRef = React.useCallback((node: HTMLDivElement | null) => {
+    if (!node) {
+      setDrawerPadding(0);
+      return;
+    }
+    if (node !== null && node.firstChild) {
+      const drawerElement = node.firstChild as HTMLElement;
+      setDrawerPadding(drawerElement.clientWidth);
+    }
+  }, []);
 
-    if (!node) {
-
-      setDrawerPadding(0);
-
-      return;
-
-    }
-
-    if (node !== null && node.firstChild) {
-
-      const drawerElement = node.firstChild as HTMLElement;
-
-      setDrawerPadding(drawerElement.clientWidth);
-
-    }
-
-  }, []);
   const toggleDrawer = () => {
-    setOpen(!open);
+    setMenuOpen(!menuOpen);
+  };
+
+  const onNavigationEntryClicked = (path: string) => {
+    setMenuOpen(false);
+    navigate(path);
   };
 
   return (
-    <ThemeProvider theme={mdTheme}>
-      <Box sx={{ display: 'flex' }}>
-        <CssBaseline />
-        <AppBar position="absolute" open={open}>
-          <Toolbar
-            sx={{
-              pr: '24px', // keep right padding when drawer closed
-            }}
-          >
-            <IconButton
-              edge="start"
-              color="inherit"
-              aria-label="open drawer"
-              onClick={toggleDrawer}
-              sx={{
-                marginRight: '36px',
-                ...(open && { display: 'none' }),
-              }}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Typography
-              component="h1"
-              variant="h6"
-              color="inherit"
-              noWrap
-              sx={{ flexGrow: 1 }}
-            >
-              Workflow Dashboard
-            </Typography>
+    <React.Fragment>
+      <React.Fragment>
+        <AppBar position="fixed">
+          <Toolbar variant={isMobile ? 'dense' : 'regular'}>
+            <Grid container justifyContent="space-between" alignItems="center">
+              <Grid item>
+                <IconButton onClick={() => setMenuOpen(!menuOpen)} edge="start" color="inherit" size="large">
+                  <MenuIcon />
+                </IconButton>
+              </Grid>
+              <Grid item>
+                <Typography component="h1" variant="h6" color="inherit">
+                  Workflow Dashboard
+                </Typography>
+              </Grid>
+            </Grid>
           </Toolbar>
         </AppBar>
-        <Drawer variant="permanent" open={open}>
-          <Toolbar
+
+        <Drawer
+          ref={measuredDrawerRef}
+          variant={isMobile ? 'temporary' : 'permanent'}
+          open={menuOpen}
+          onClose={() => setMenuOpen(false)}
+        >
+          <List
             sx={{
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'flex-end',
-              px: [1],
+              paddingTop: 0,
+              a: {
+                textDecoration: 'none',
+                color: 'black',
+              },
             }}
           >
-            <IconButton onClick={toggleDrawer}>
-              <ChevronLeftIcon />
+            <IconButton sx={{ minHeight: isMobile ? '56px' : '64px', boxSizing: 'border-box' }}>
+              <img src="logo192.png" alt="Logo" height={36} />
             </IconButton>
-          </Toolbar>
-          <Divider />
-          <List component="nav">
-            {mainListItems}
-            <Divider sx={{ my: 1 }} />
-            {secondaryListItems}
+            <Divider />
+            <React.Fragment>
+              <ListItemButton selected={location.pathname === '/'} onClick={() => onNavigationEntryClicked('/')}>
+                <ListItemIcon>
+                  <DashboardIcon />
+                </ListItemIcon>
+                <ListItemText primary="Dashboard" />
+              </ListItemButton>
+
+              <ListItemButton
+                selected={location.pathname === '/workflows'}
+                onClick={() => onNavigationEntryClicked('/workflows')}
+              >
+                <ListItemIcon>
+                  <AltRouteIcon />
+                </ListItemIcon>
+                <ListItemText primary="Workflows" />
+              </ListItemButton>
+            </React.Fragment>
           </List>
         </Drawer>
-      </Box>
-      <main style={{ marginTop: '7rem', paddingLeft: '${drawerPadding}px' }}>
+      </React.Fragment>
+      <main style={{ marginTop: '7rem', paddingLeft: `${drawerPadding}px` }}>
         <Routes>
-          <Route path="/" element= {<Dashboard />} />
+          <Route path="/" element={<Dashboard />} />
           <Route path="/workflows" element={<div>Hi blablablablablalbalblabalbla</div>} />
+          <Route element={<div>Not Found</div>} />
         </Routes>
       </main>
-    </ThemeProvider>
+    </React.Fragment>
   );
 }
 
-export default NavigationDrawer
+export default NavigationDrawer;
