+++ frontend/src/Pages/Dashboard/Runners.tsx
@@ -1,25 +1,16 @@
 import * as React from 'react';
-import Link from '@mui/material/Link';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
 import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
-import Title from './Title';
-import { useEffect, useState } from "react";
-import TableContainer from '@mui/material/TableContainer';
+import Title from '../../components/Title';
+import { useEffect, useState } from 'react';
 import TableFooter from '@mui/material/TableFooter';
 import TablePagination from '@mui/material/TablePagination';
-import Paper from '@mui/material/Paper';
-import IconButton from '@mui/material/IconButton';
-import FirstPageIcon from '@mui/icons-material/FirstPage';
-import KeyboardArrowLeft from '@mui/icons-material/KeyboardArrowLeft';
-import KeyboardArrowRight from '@mui/icons-material/KeyboardArrowRight';
-import LastPageIcon from '@mui/icons-material/LastPage';
 import { CheckCircle, Warning, Schedule } from '@mui/icons-material';
 import { ListItemIcon } from '@mui/material';
 import Typography from '@mui/material/Typography';
-import Grid from '@mui/material/Grid';
 import Stack from '@mui/material/Stack';
 
 async function getRunners() {
@@ -52,9 +43,7 @@ export default function Runners() {
     setPage(newPage);
   };
 
-  const handleRowsPerPageChange = (
-    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
-  ) => {
+  const handleRowsPerPageChange = (event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
     setRowsPerPage(parseInt(event.target.value, 10));
     setPage(0);
   };
@@ -88,29 +77,29 @@ export default function Runners() {
   return (
     <React.Fragment>
       <Stack direction="row" justifyContent="space-between" alignItems="flex-start" spacing={2}>
-          <Title>Runners</Title>
-          <React.Fragment>
-            <Typography variant="heading" color="textSecondary" style={{ display: 'flex', alignItems: 'center' }}>
-              <TableCell>
-                <ListItemIcon>
-                  <Warning style={{ color: 'red' }} />
-                </ListItemIcon>
-                Offline
-              </TableCell>
-              <TableCell>
-                <ListItemIcon>
-                  <CheckCircle style={{ color: 'green'}} />
-                </ListItemIcon>
-                Online / Ready
-              </TableCell>
-              <TableCell>
-                <ListItemIcon>
-                  <Schedule style={{ color: 'primary' }} />
-                </ListItemIcon>
-                Online / Busy
-              </TableCell>
-            </Typography>
-          </React.Fragment>
+        <Title>Runners</Title>
+        <React.Fragment>
+          <Typography variant="h1" color="textSecondary" style={{ display: 'flex', alignItems: 'center' }}>
+            <TableCell>
+              <ListItemIcon>
+                <Warning style={{ color: 'red' }} />
+              </ListItemIcon>
+              Offline
+            </TableCell>
+            <TableCell>
+              <ListItemIcon>
+                <CheckCircle style={{ color: 'green' }} />
+              </ListItemIcon>
+              Online / Ready
+            </TableCell>
+            <TableCell>
+              <ListItemIcon>
+                <Schedule style={{ color: 'primary' }} />
+              </ListItemIcon>
+              Online / Busy
+            </TableCell>
+          </Typography>
+        </React.Fragment>
       </Stack>
       <Table size="small">
         <TableHead>
@@ -131,16 +120,13 @@ export default function Runners() {
                   {getStatusIcon(runner.status, runner.busy)}
                 </ListItemIcon>
               </TableCell>
-              <TableCell>
-                {runner.labels.map((label: { name: string }) => label.name).join(', ')}
-              </TableCell>
+              <TableCell>{runner.labels.map((label: { name: string }) => label.name).join(', ')}</TableCell>
             </TableRow>
           ))}
         </TableBody>
         <TableFooter>
           <TableRow>
-          <TableBody>
-          </TableBody>
+            <TableBody></TableBody>
             <TablePagination
               rowsPerPageOptions={[5, 10, 25, { label: 'All', value: -1 }]}
               colSpan={5}
