--- frontend/src/listItems.tsx
@@ -1,42 +1,0 @@
-import * as React from 'react';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
-import ListSubheader from '@mui/material/ListSubheader';
-import DashboardIcon from '@mui/icons-material/Dashboard';
-import AltRouteIcon from '@mui/icons-material/AltRoute';
-import PeopleIcon from '@mui/icons-material/People';
-import BarChartIcon from '@mui/icons-material/BarChart';
-import DateRangeIcon from '@mui/icons-material/DateRange';
-import LayersIcon from '@mui/icons-material/Layers';
-import AssignmentIcon from '@mui/icons-material/Assignment';
-import Grid from '@mui/material/Grid';
-import Paper from '@mui/material/Paper';
-
-
-export const mainListItems = (
-  <React.Fragment>
-    <ListItemButton>
-      <ListItemIcon>
-        <DashboardIcon />
-      </ListItemIcon>
-      <ListItemText>
-         <a href={`/`}>Dashboard</a>
-      </ListItemText>
-    </ListItemButton>
-    <ListItemButton>
-      <ListItemIcon>
-        <AltRouteIcon />
-      </ListItemIcon>
-      <ListItemText>
-         <a href={`/workflows`}>Workflows</a>
-      </ListItemText>
-    </ListItemButton>
-  </React.Fragment>
-);
-
-export const secondaryListItems = (
-  <React.Fragment>
-
-  </React.Fragment>
-);
